+++ app/src/main/java/de/vier_bier/habpanelviewer/reporting/BrightnessMonitor.java
@@ -16,7 +16,7 @@ public class BrightnessMonitor extends AbstractAveragingDeviceMonitor {
     }
 
     @Override
-    String getInfoString(Integer value, String item, String state) {
+    String getInfoString(Float value, String item, String state) {
         return mCtx.getString(R.string.brightness, mValue, mSensorItem, mSensorState);
     }
 }
