+++ app/src/main/java/de/vier_bier/habpanelviewer/reporting/TemperatureMonitor.java
@@ -17,7 +17,7 @@ public class TemperatureMonitor extends AbstractAveragingDeviceMonitor {
 
 
     @Override
-    String getInfoString(Integer value, String item, String state) {
+    String getInfoString(Float value, String item, String state) {
         return mCtx.getString(R.string.temperature, mValue, mSensorItem, mSensorState);
     }
 }
